refactor: `getFilterObjectPath`
