+++ app/src/components/features/rules/RulePairs/Filters/actions/deleteObjectAtPath.js
@@ -1,4 +1,5 @@
-import { getFilterObjectPath } from "../../index";
+import { getFilterObjectPath } from "utils/rules/getFilterObjectPath";
+
 const omit = require("lodash/omit");
 
 const deleteObjectAtPath = (currentlySelectedRuleData, setCurrentlySelectedRule, targetPath, pairIndex, dispatch) => {
