+++ app/src/utils/rules/getFilterObjectPath.js
@@ -1,0 +1,8 @@
+export const getFilterObjectPath = (objectPath, filterIndex = 0) => {
+  if (!objectPath.includes("filters")) {
+    return objectPath;
+  }
+  const splitObjectPath = objectPath.split(".");
+  splitObjectPath[1] = splitObjectPath[1] + `[${filterIndex}]`;
+  return splitObjectPath.join(".");
+};
